// backend/src/modules/production-orders/dto/update-production-order-tablet-thickness-check.dto.ts
export class UpdateProductionOrderTabletThicknessCheckDto {
  requirement?: string | null;
  dosage_form_stage?: string | null;
  unit_1_thickness?: number | string | null;
  unit_2_thickness?: number | string | null;
  unit_3_thickness?: number | string | null;
  unit_4_thickness?: number | string | null;
  unit_5_thickness?: number | string | null;
  unit_6_thickness?: number | string | null;
  unit_7_thickness?: number | string | null;
  unit_8_thickness?: number | string | null;
  unit_9_thickness?: number | string | null;
  unit_10_thickness?: number | string | null;
  unit?: string | null;
}